Move helper types to separate file
