+++ src/Http/Http.Extensions/gen/StaticRouteHandlerModel/Model/ConstructorParameter.cs
@@ -1,0 +1,7 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+using Microsoft.CodeAnalysis;
+
+namespace Microsoft.AspNetCore.Http.RequestDelegateGenerator.StaticRouteHandlerModel;
+
+internal record ConstructorParameter(IPropertySymbol Property, IParameterSymbol? Parameter);
